Rename _cpp_bindings to _pywavemap_bindings s.t. name is self explaining
