+++ library/python/src/pywavemap/__init__.py
@@ -1,16 +1,16 @@
 # Use module doc string as pkg doc string
-from ._cpp_bindings import __doc__
+from ._pywavemap_bindings import __doc__
 
 # Binding types
-from ._cpp_bindings import OctreeIndex
-from ._cpp_bindings import (Rotation, Pose, Pointcloud, PosedPointcloud, Image,
-                            PosedImage)
-from ._cpp_bindings import (Map, HashedWaveletOctree,
-                            HashedChunkedWaveletOctree)
-from ._cpp_bindings import Pipeline
+from ._pywavemap_bindings import OctreeIndex
+from ._pywavemap_bindings import (Rotation, Pose, Pointcloud, PosedPointcloud,
+                                  Image, PosedImage)
+from ._pywavemap_bindings import (Map, HashedWaveletOctree,
+                                  HashedChunkedWaveletOctree)
+from ._pywavemap_bindings import Pipeline
 
 # Binding submodules
-from ._cpp_bindings import logging, param
+from ._pywavemap_bindings import logging, param
 
 # Regular modules
 from . import convert
